Code Documentation/requirements.txt
openai==1.14.3
streamlit==1.32.2
dotenv-python==0.0.1
mysql-connector-python==8.3.0
pandas==2.2.1